Added type conversion to model

+++ parllel/jax/models.py
@@ -21,7 +21,7 @@ class ActorCriticModel(nn.Module):
 
     @nn.compact
     def __call__(self, input):
-        x = input
+        x = input.astype(jnp.float32)
         for size in self.actor_hidden_sizes:
             x = nn.Dense(features=size)(x)
             x = nn.relu(x)
